create requirement view title factory
